Speech bubble, name tag and collision detection
Fixed speech bubble's position. Now it doesn't follow the player but instead npc wherever it moves.
Remade the name tag system to work for all entities. Created a system to correctly center the name tag based on the size of name string.
Added a collision detection system that won't allow an entity to move if it hits a collision.